Move inputs file to a separate folder.
This patch moves input related files to under inputs folder.
