+++ wayland.gyp
@@ -43,10 +43,6 @@
         '../..',
       ],
       'sources': [
-        'wayland/egl/egl_window.cc',
-        'wayland/egl/egl_window.h',
-        'wayland/cursor.cc',
-        'wayland/cursor.h',
         'wayland/display.cc',
         'wayland/display.h',
         'wayland/dispatcher.cc',
@@ -54,16 +50,6 @@
         'wayland/global.h',
         'wayland/input_device.cc',
         'wayland/input_device.h',
-        'wayland/input_method_event_filter.cc',
-        'wayland/input_method_event_filter.h',
-        "wayland/kbd_conversion.cc",
-        "wayland/kbd_conversion.h",
-        "wayland/keyboard.cc",
-        "wayland/keyboard.h",
-        "wayland/egl/loader.h",
-        "wayland/egl/loader.cc",
-        "wayland/pointer.cc",
-        "wayland/pointer.h",
         'wayland/screen.cc',
         'wayland/screen.h',
         'wayland/surface.h',
@@ -72,6 +58,20 @@
         'wayland/shell_surface.cc',
         'wayland/window.cc',
         'wayland/window.h',
+        'wayland/egl/egl_window.cc',
+        'wayland/egl/egl_window.h',
+        'wayland/egl/loader.h',
+        'wayland/egl/loader.cc',
+        'wayland/inputs/cursor.cc',
+        'wayland/inputs/cursor.h',
+        'wayland/inputs/input_method_event_filter.cc',
+        'wayland/inputs/input_method_event_filter.h',
+        'wayland/inputs/kbd_conversion.cc',
+        'wayland/inputs/kbd_conversion.h',
+        'wayland/inputs/keyboard.cc',
+        'wayland/inputs/keyboard.h',
+        'wayland/inputs/pointer.cc',
+        'wayland/inputs/pointer.h',
       ],
     },
   ]
