+++ wayland/inputs/OWNERS
@@ -1,0 +1,2 @@
+kalyan.kondapally@intel.com
+tiago.vignatti@intel.com
